Conflicts:
	WeatherBalloon/WeatherBalloon.atsuo
	WeatherBalloon/WeatherBalloonSensor/Drivers/BMP180.c
	WeatherBalloon/WeatherBalloonSensor/WeatherBalloonSensor.c
	WeatherBalloon/WeatherBalloonSensor/WeatherBalloonSensor.cproj

+++ Hello_LM3115_ATMega32/Hello_LM3115_ATMega32/HeaderFiles/BMP180.h
@@ -13,7 +13,7 @@
 
 void BMP180_Init();
 unsigned char BMP180_GetDeviceId();
-unsigned long int BMP180_GetTemperature();
+short BMP180_GetTemperature();
 
 //unsigned char BMP180_RegRead(unsigned char[] RetVal, unsigned char reg, unsigned int NumOfBytes);
 //void BMP180_RegWrite(unsigned char reg, unsigned char val);
